+++ src/mqt/predictor/rl/__init__.py
@@ -14,12 +14,11 @@ See README.md or go to https://github.com/munich-quantum-toolkit/predictor for m
 
 from __future__ import annotations
 
-from mqt.predictor.rl.helper import qcompile
-from mqt.predictor.rl.predictor import Predictor
+from mqt.predictor.rl.predictor import Predictor, rl_compile
 from mqt.predictor.rl.predictorenv import PredictorEnv
 
 __all__ = [
     "Predictor",
     "PredictorEnv",
-    "qcompile",
+    "rl_compile",
 ]