+++ CHANGELOG.md
@@ -11,7 +11,10 @@ This project adheres to [Semantic Versioning], with the exception that minor rel
 
 ### Changed
 
+- ✨ Improved the ML part and its usability ([#403]) ([**@nquetschlich**])
+- 📝 Migrated the documentation from .rst to .md files ([#403]) ([**@nquetschlich**])
 - ✨ Improve RL action handling by using dataclasses ([#401]) ([**@nquetschlich**])
+- ✨ Support MQT Bench v2 and use Qiskit's Target to represent quantum devices ([#393]) ([**@nquetschlich**])
 
 ## [2.2.0] - 2025-02-02
 
@@ -23,6 +26,10 @@ _📚 Refer to the [GitHub Release Notes](https://github.com/munich-quantum-tool
 [2.2.0]: https://github.com/munich-quantum-toolkit/predictor/releases/tag/v2.2.0
 
 <!-- PR links -->
+
+[#403]: https://github.com/munich-quantum-toolkit/predictor/pull/403
+[#401]: https://github.com/munich-quantum-toolkit/predictor/pull/401
+[#393]: https://github.com/munich-quantum-toolkit/predictor/pull/393
 
 <!-- Contributor -->